Bulk_email/requirements.txt
fastapi
slowapi
aiosmtplib
pydantic
python-dotenv
python-multipart
email-validator
uvicorn